mv func to cpp_prefix
